Fix multiple important bugs. LAv2 works in CPU-only mode!  Results look pretty much the same as BLAv1.

+++ LAReference.h
@@ -61,7 +61,7 @@ public:
 
     ATInfo AT;
 
-    int LAStageCount;
+    size_t LAStageCount;
 
     bool isValid;
